support for yadage dialects
This PR extends the loader to load workflows via a number
of 'registered' dialects that all are able to compile
to JSON that validates against the yadage JSON schema

The motivation to add this is that we already had a number
of 'massaging' to the loaded JSON/YAML in order to not have
the user specify absolutely everything (e.g. respect defaults
set in the JSON schema) and over time we added some
non-trivial syntactic sugar (e.g. stage dependencies evoled from
lists to objects, and in order to be backwards compatible we
now treat lists as syntactic sugar and expand them to full
objects during load-time)

We now treat the full JSON as a verbose but hopefully stable
intermediate representation. The job of 'dialects' is to
compile whatever input they are given into a JSON document
that validates against the JSON schema for yadage workflows.

An example of such a dialect is the 'dagdown' prototype

https://github.com/lukasheinrich/dagdown

which will be moved to

https://github.com/yadage/yadage-langs

that represents a much more concise but more limited mini-language
for describing workflows.

Separating the input format from an intermediate represenation also
allows us to experiment more freely with the input format as long
as we can compile to the IR, while more core tools (such
as visualizers etc keep using themore table IR.
