+++ yadageschemas/dialects/__init__.py
@@ -1,0 +1,3 @@
+from ..utils import handler_decorator
+
+handlers, dialect  = handler_decorator()
